backend/requirements.txt
fastapi
uvicorn
gunicorn
supabase
python-dotenv
pydantic
pytest
httpx
openai
email-validator
